+++ src/platform/linux/x11/window.rs
@@ -332,6 +332,8 @@ impl Window2 {
         let mut find = default;
         for monitor in monitors {
             let (mx, my) = monitor.get_position();
+            let mx = mx as u32;
+            let my = my as u32;
             let (mw, mh) = monitor.get_dimensions();
             let (mxo, myo) = (mx+mw-1, my+mh-1);
             let (ox, oy) = (cmp::max(wx, mx), cmp::max(wy, my));
@@ -473,7 +475,7 @@ impl Window2 {
 
     #[inline]
     pub fn get_inner_size(&self) -> Option<(u32, u32)> {
-        self.get_geometry().map(|(_, _, w, h, _)| ((w as f32 / self.hidpi_factor()) as u32, (h as f32 / self.hidpi_factor()) as u32))
+        self.get_geometry().map(|(_, _, w, h, _)| (w, h))
     }
 
     #[inline]